Runner基础接口

+++ Runner_Entity/src/main/java/com/runner/entity/pojo/GainMuscle.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class GainMuscle {
 
-  private long gainMuscleId;
-  private long gainMusclePart;
+  private Integer gainMuscleId;
+  private Integer gainMusclePart;
   private String partStatus;
   private String improveStatus;
-  private long planInfoId;
+  private Integer planInfoId;
 
 }
